requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
scipy>=1.9.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
joblib>=1.2.0
pyyaml>=6.0
pytest>=7.0.0
umap-learn>=0.5.0
aiofiles>=23.0.0
